+++ x-pack/plugin/core/src/test/java/org/elasticsearch/xpack/core/dataframe/transforms/MockDeprecatedQueryBuilder.java
@@ -4,7 +4,7 @@
  * you may not use this file except in compliance with the Elastic License.
  */
 
-package org.elasticsearch.xpack.core.dataframe.transform;
+package org.elasticsearch.xpack.core.dataframe.transforms;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.lucene.search.Query;
